+++ client/src/App.js
@@ -10,6 +10,7 @@ function App() {
     <>
       <Router>
         <AuthContextProvider>
+          <NavBar />
           <AppRouter />
           <Footer />
         </AuthContextProvider>
